+++ src/index.js
@@ -2,6 +2,10 @@ var ip = require('ip')
 var os = require('os')
 var net = require('net')
 var cp = require('mz/child_process')
+
+var parseLinux = require('./parser/linux')
+var parseRow = require('./parser')
+
 var servers = getServers()
 var lock = {}
 
@@ -21,7 +25,8 @@ module.exports = function findLocalDevices (address) {
 
 /**
  * Gets the current list of possible servers in the local networks.
- * TODO: mock in tests
+ *
+ * TODO: mock in tests (eg. also offline case)
  */
 function getServers () {
   var interfaces = os.networkInterfaces()
@@ -56,7 +61,7 @@ function pingServers () {
 function pingServer (address) {
   return new Promise(function (resolve) {
     // https://stackoverflow.com/a/51986121/1238150
-    const socket = new net.Socket()
+    var socket = new net.Socket()
     socket.setTimeout(1000, close)
     socket.connect(80, address, close)
     socket.once('error', close)
@@ -83,16 +88,19 @@ function parseAll (data) {
     return []
   }
 
-  // detect OS
-  const platform = process.platform
-  if (platform.includes('linux')) {
-    return parseLinux(data, true)
+  if (process.platform.includes('linux')) {
+    var rows = data[0].split('\n')
+    return rows.map(function (row) {
+      return parseLinux(row)
+    }).filter(Boolean)
   }
 
   return data[0]
     .trim()
     .split('\n')
-    .map(parseRow)
+    .map(function (row) {
+      return parseRow(row, servers)
+    })
     .filter(Boolean)
 }
 
@@ -111,93 +119,18 @@ function parseOne (data) {
     return
   }
 
-  // detect OS
-  const platform = process.platform
-  if (platform.includes('linux')) {
-    return parseLinux(data)
-  }
-
-  return parseRow(data[0])
-}
-
-/**
- * Parses each row in the arp table into { name, ip, mac } on MACOSX.
- */
-function parseRow (row) {
-  // Parse name.
-  var nameStart = 0
-  var nameEnd = row.indexOf('(') - 1
-  var name = row.slice(nameStart, nameEnd)
-
-  // Parse ip.
-  var ipStart = nameEnd + 2
-  var ipEnd = row.indexOf(')', ipStart)
-  var ipAddress = row.slice(ipStart, ipEnd)
-  // Only resolve external ips.
-  if (!~servers.indexOf(ipAddress)) {
-    return
-  }
-
-  // Parse mac
-  var macStart = row.indexOf(' at ', ipEnd) + 4
-  var macEnd = row.indexOf(' on ', macStart)
-  var macAddress = row.slice(macStart, macEnd)
-  // Ignore unresolved hosts.
-  if (macAddress === '(incomplete)') {
-    return
-  }
-  // Format for always 2 digits
-  macAddress = macAddress
-    .replace(/^.:/, '0$&')
-    .replace(/:.(?=:|$)/g, ':0X$&')
-    .replace(/X:/g, '')
-
-  return {
-    name: name,
-    ip: ipAddress,
-    mac: macAddress
-  }
-}
-
-/**
- * Parses each row in the arp table into { name, ip, mac } on LINUX.
- */
-function parseLinux (data, parseMultiple = false) {
-  console.log('data', data); // eslint-disable-line 
-  // ignore unresolved hosts (can happen when parseOne returns only one unresolved host)
-  if (data.indexOf('no entry') >= 0) {
-    return
-  }
-
-  // partial inspired by https://github.com/goliatone/arpscan/blob/master/lib/arpscanner.js
-  const out = []
-  const rows = data.split('\n').slice(parseMultiple ? 0 : 1)
-
-  // Parses each row in the arp table into { name, ip, mac }.
-  rows.forEach(row => {
-    // Ignore unresolved hosts.
-    if (row === '' || row.indexOf('incomplete') >= 0) {
+  if (process.platform.includes('linux')) {
+    // ignore unresolved hosts (can happen when parseOne returns only one unresolved host)
+    if (data[0].indexOf('no entry') >= 0) {
       return
     }
-    const chunk = row.split(' ').filter(Boolean)
 
-    if (parseMultiple) {
-      return out.push({
-        name: chunk[0],
-        ip: chunk[1].match(/\((.*)\)/)[1],
-        mac: chunk[3]
-      })
-    }
+    // remove first row (containing "headlines")
+    var rows = data[0].split('\n').slice(1)[0]
+    return parseLinux(rows, true)
+  }
 
-    out.push({
-      name: '?', // does not provide a device name
-      ip: chunk[0],
-      mac: chunk[2]
-    })
-  })
-
-  const result = out.filter(Boolean)
-  return result.length ? result : undefined;
+  return parseRow(data[0], servers)
 }
 
 /**